micro: share session query parsing
