+++ frontend/src/app/api/agent/sessions/all/route.ts
@@ -3,6 +3,7 @@ import { existsSync, statSync } from "node:fs";
 import { listProjectsFromStore } from "@/lib/agent/projects-store";
 import { listSessions, type SessionSummary } from "@/lib/agent/sessions-store";
 import { listArchivedSessionMetadata } from "@/lib/agent/session-metadata-store";
+import { archiveQueryOptions, parseRelativeSince } from "../session-query";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
@@ -16,35 +17,10 @@ export type AggregatedSession = SessionSummary & {
   projectPath: string;
 };
 
-function parseSince(value: string | null): Date | null {
-  if (!value) return null;
-  const match = value.match(/^(\d+)([dhm])$/);
-  if (!match) return null;
-  const amount = Number(match[1]);
-  if (!Number.isFinite(amount) || amount <= 0) return null;
-  const unit = match[2];
-  const multiplier = unit === "d" ? 86_400_000 : unit === "h" ? 3_600_000 : 60_000;
-  return new Date(Date.now() - amount * multiplier);
-}
-
-function archiveOptions(searchParams: URLSearchParams): {
-  includeArchived?: boolean;
-  archivedOnly?: boolean;
-} {
-  const archived = searchParams.get("archived")?.toLowerCase();
-  const includeArchived = searchParams.get("includeArchived")?.toLowerCase();
-  return {
-    ...(includeArchived === "1" || includeArchived === "true" ? { includeArchived: true } : {}),
-    ...(archived === "1" || archived === "true" || archived === "only"
-      ? { archivedOnly: true, includeArchived: true }
-      : {}),
-  };
-}
-
 export async function GET(request: NextRequest) {
   const sinceParam = request.nextUrl.searchParams.get("since");
-  const since = parseSince(sinceParam) ?? undefined;
-  const archive = archiveOptions(request.nextUrl.searchParams);
+  const since = parseRelativeSince(sinceParam) ?? undefined;
+  const archive = archiveQueryOptions(request.nextUrl.searchParams);
   const projects = listProjectsFromStore();
   const aggregated: AggregatedSession[] = [];
   const seenIds = new Set<string>();
